Config: Add browser-safe `./schema` entry point
