+++ javascript/packages/config/rolldown.config.mjs
@@ -24,4 +24,28 @@ export default [
     external: ["yaml", "fs", "path", "picomatch", "tinyglobby"],
     plugins: [yaml()],
   },
+
+  {
+    input: "src/config-schema.ts",
+    output: {
+      file: "dist/herb-config-schema.esm.js",
+      format: "esm",
+      sourcemap: true,
+      codeSplitting: false,
+    },
+    external: ["yaml", "picomatch"],
+    plugins: [yaml()],
+  },
+
+  {
+    input: "src/config-schema.ts",
+    output: {
+      file: "dist/herb-config-schema.cjs",
+      format: "cjs",
+      sourcemap: true,
+      codeSplitting: false,
+    },
+    external: ["yaml", "picomatch"],
+    plugins: [yaml()],
+  },
 ]
